Revision: Revisando códigos.

+++ src/controllers/groupMuscleController.ts
@@ -1,27 +1,70 @@
 import type { FastifyRequest, FastifyReply } from 'fastify';
 import { prisma } from "../lib/prisma.js";
 import type { GroupMuscle } from "../types/groupMuscle.js";
+import { groupMuscleSchema } from '../schemas/groupMuscleSchemas.js';
 
 const createGroupMuscle = async (request: FastifyRequest, reply: FastifyReply) => {
   try {
+    const validateGroupMuscleInformations = groupMuscleSchema.safeParse(request.body as GroupMuscle);
+    if (!validateGroupMuscleInformations.success) return reply.status(400).send({ error: "Invalid group muscle informations." });
+
     const { name, description } = request.body as GroupMuscle;
 
-    const gm = await prisma.groupMuscle.create({
-      data: { name, description }
+    const groupMuscle = await prisma.groupMuscle.create({
+      data: {
+        name: name,
+        description: description
+      }
     });
 
-    return reply.status(201).send(gm);
+    return {
+      id: groupMuscle.id,
+      name: groupMuscle.name,
+      description: groupMuscle.description,
+      createdAt: groupMuscle.createdAt,
+      updatedAt: groupMuscle.updatedAt
+    };
   } catch (error) {
     return reply.status(500).send({ error, message: "Unable to create group muscle." });
   }
 }
 
+const updateGroupMuscle = async (request: FastifyRequest, reply: FastifyReply) => {
+  const { id } = request.params as { id: string };
+
+  const validateGroupMuscleInformations = groupMuscleSchema.partial().safeParse(request.body as GroupMuscle);
+  if (!validateGroupMuscleInformations.success) return reply.status(400).send({ error: "Invalid group muscle informations." })
+
+  const { name, description } = validateGroupMuscleInformations.data;
+
+  const updateData: {
+    name?: string;
+    description?: string;
+  } = {};
+
+  if (name) updateData.name = name;
+
+  if (description) updateData.description = description;
+
+  try {
+    const updated = await prisma.groupMuscle.update({
+      where: { id },
+      data: updateData
+    });
+
+    return updated;
+  } catch (error) {
+    return reply.status(404).send({ error, message: "Unable to update the group muscle. It may not exist." });
+  }
+}
+
 const getAllGroupMuscles = async (_request: FastifyRequest, reply: FastifyReply) => {
   try {
-    const gms = await prisma.groupMuscle.findMany();
-    return reply.status(200).send(gms);
+    const groupMuscles = await prisma.groupMuscle.findMany();
+
+    return groupMuscles;
   } catch (error) {
-    return reply.status(500).send({ error, message: "Unable to fetch group muscles." });
+    return reply.status(404).send({ error, message: "Group muscles not found." });
   }
 }
 
@@ -36,26 +79,14 @@ const getGroupMuscleById = async (request: FastifyRequest, reply: FastifyReply) 
   }
 }
 
-const updateGroupMuscle = async (request: FastifyRequest, reply: FastifyReply) => {
-  try {
-    const { id } = request.params as { id: string };
-    const { name, description } = request.body as GroupMuscle;
-
-    const updated = await prisma.groupMuscle.update({
-      where: { id },
-      data: { name, description }
-    });
-
-    return reply.status(200).send(updated);
-  } catch (error) {
-    return reply.status(404).send({ error, message: "Unable to update the group muscle. It may not exist." });
-  }
-}
-
 const deleteGroupMuscle = async (request: FastifyRequest, reply: FastifyReply) => {
+  const { id } = request.params as { id: string };
+
+  if (!id) return reply.status(404).send({ error: "Group muscle not found." })
+
   try {
-    const { id } = request.params as { id: string };
-    await prisma.groupMuscle.delete({ where: { id }});
+    await prisma.groupMuscle.delete({ where: { id } });
+
     return reply.status(200).send({ message: "Group muscle deleted successfully!" });
   } catch (error) {
     return reply.status(404).send({ error, message: "Unable to delete the group muscle. An error occurred." });
